Change from a ring undo buffer to a linear array
With this change the undo features work as expected for the positioning.
The only flaw is that because the history is strictly linear we have to
discard history if the user starts a new branch by making new changes
while back in time. I think there must be a better way to handle it than
what I do, which is to completely clear the history and then add back in
the state right before the new change and the new change.